+++ tests/component/SequenceDisplayPreview.test.ts
@@ -3,10 +3,10 @@ import { describe, test, vi } from 'vitest';
 import SequenceDisplayPreview from '../../../frontend/src/components/SequenceDisplayPreview.vue';
 import PopupComponent from '../../../frontend/src/components/PopupComponent.vue';
 import ButtonComponent from '../../../frontend/src/components/ButtonComponent.vue';
-import { routerMock } from '../router_mock.setup';
+import { routerMock } from './router_mock.setup';
 import TagSequencePopupTab from '../../../frontend/src/components/sequenceOverviewPopUpTabs/TagSequencePopupTab.vue';
 import { defineComponent } from 'vue';
-import { AccountRestInterface } from '../../../frontend/src/restInterfaces/AccountRestInterface';
+import { AccountRestInterface } from '../../frontend/src/restInterfaces/AccountRestInterface';
 
 describe('SequenceDisplayPreview', () => {
     const baseSequence = {
